requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
xlsxwriter==3.1.9
pyecharts==2.0.4
werkzeug>=2.0.0
gunicorn>=20.1.0